+++ packages/liveblocks-react/src/index.ts
@@ -75,7 +75,9 @@ export {
   useUpdateMyPresence,
   useUpdateRoomSubscriptionSettings,
   useHistoryVersionData,
+  useHistoryVersionStorageData,
   useHistoryVersionYjsData,
+  useDeleteHistoryVersion,
 } from "./room";
 
 // Export the classic (non-Suspense) versions of our hooks